Improve typing of user session (#1258)
