+++ public/locales/bg/profile.json
@@ -3,7 +3,7 @@
   "campaigns": "Подкрепени кампании",
   "personalInfo": {
     "index": "Лична информация",
-    "login": "Login информация:",
+    "login": "Данни за вход:",
     "email": "Email адрес:",
     "edit": "Редактирай",
     "password": "Парола:",
