+++ src/components/layout/nav/AdminMenu.tsx
@@ -67,7 +67,7 @@ export default function AdminMenu() {
     return null
   }
 
-  const title = `${session.user.name}\n(${session.user.email})`
+  const title = `${session?.user?.name}\n(${session?.user?.email})`
   const lettersAvatar = `${session.user?.given_name.charAt(0)}${session.user?.family_name.charAt(
     0,
   )}`.toUpperCase()
